feat: Add tests for embedding classifier using remote provider
